requirements.txt
Django>=5.2.1,<5.9.0
djangorestframework>=3.15.1,<3.20.0
django-redis>=5.4.0,<6.0
redis>=4.6.0,<4.9
drf-spectacular>=0.27.0,<0.28.0
psycopg2-binary==2.9.9
requests>=2.31.0,<3.0
mongoengine>=0.28.0,<1.0
pymongo>=4.6.1,<5.0
sqlparse>=0.3.1
celery>=5.5.2,<6.0
flower>=2.0.1,<3.0
python-stdnum>=2.1.0,<3.0